+++ beacon-chain/rpc/prysm/beacon/handlers.go
@@ -1,17 +1,23 @@
 package beacon
 
 import (
+	"encoding/json"
 	"fmt"
+	"io"
 	"log"
 	"net/http"
 	"strconv"
 
 	"github.com/ethereum/go-ethereum/common/hexutil"
+	"github.com/pkg/errors"
 	"github.com/prysmaticlabs/prysm/v5/api/server/structs"
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/core/helpers"
+	"github.com/prysmaticlabs/prysm/v5/beacon-chain/rpc/core"
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/rpc/eth/shared"
 	"github.com/prysmaticlabs/prysm/v5/config/params"
+	"github.com/prysmaticlabs/prysm/v5/consensus-types/primitives"
 	"github.com/prysmaticlabs/prysm/v5/network/httputil"
+	ethpb "github.com/prysmaticlabs/prysm/v5/proto/prysm/v1alpha1"
 	"github.com/prysmaticlabs/prysm/v5/time/slots"
 	"go.opencensus.io/trace"
 )
@@ -68,4 +74,87 @@ func (s *Server) GetWeakSubjectivity(w http.ResponseWriter, r *http.Request) {
 		},
 	}
 	httputil.WriteJson(w, resp)
+}
+
+// GetIndividualVotes returns a list of validators individual vote status of a given epoch.
+func (s *Server) GetIndividualVotes(w http.ResponseWriter, r *http.Request) {
+	ctx, span := trace.StartSpan(r.Context(), "validator.GetIndividualVotes")
+	defer span.End()
+
+	var req structs.GetIndividualVotesRequest
+	err := json.NewDecoder(r.Body).Decode(&req)
+	switch {
+	case errors.Is(err, io.EOF):
+		httputil.HandleError(w, "No data submitted", http.StatusBadRequest)
+		return
+	case err != nil:
+		httputil.HandleError(w, "Could not decode request body: "+err.Error(), http.StatusBadRequest)
+		return
+	}
+
+	publicKeyBytes := make([][]byte, len(req.PublicKeys))
+	for i, s := range req.PublicKeys {
+		bss := make([]byte, len(s))
+		for j, hexStr := range s {
+			bs, err := hexutil.Decode(hexStr)
+			if err != nil {
+				httputil.HandleError(w, "could not decode public keys: "+err.Error(), http.StatusBadRequest)
+				return
+			}
+			bss[j] = bs[0]
+		}
+		publicKeyBytes[i] = bss
+	}
+	epoch, err := strconv.ParseUint(req.Epoch, 10, 64)
+	if err != nil {
+		httputil.HandleError(w, "invalid epoch: "+err.Error(), http.StatusBadRequest)
+		return
+	}
+	var indices []primitives.ValidatorIndex
+	for _, i := range req.Indices {
+		u, err := strconv.ParseUint(i, 10, 64)
+		if err != nil {
+			httputil.HandleError(w, "invalid indices: "+err.Error(), http.StatusBadRequest)
+			return
+		}
+		indices = append(indices, primitives.ValidatorIndex(u))
+	}
+	votes, rpcError := s.CoreService.IndividualVotes(
+		ctx,
+		&ethpb.IndividualVotesRequest{
+			Epoch:      primitives.Epoch(epoch),
+			PublicKeys: publicKeyBytes,
+			Indices:    indices,
+		},
+	)
+
+	if rpcError != nil {
+		httputil.HandleError(w, rpcError.Err.Error(), core.ErrorReasonToHTTP(rpcError.Reason))
+		return
+	}
+	v := make([]*structs.IndividualVote, 0, len(votes.IndividualVotes))
+	for _, vote := range votes.IndividualVotes {
+		v = append(v, &structs.IndividualVote{
+			Epoch:                            fmt.Sprintf("%d", vote.Epoch),
+			PublicKey:                        hexutil.Encode(vote.PublicKey),
+			ValidatorIndex:                   fmt.Sprintf("%d", vote.ValidatorIndex),
+			IsSlashed:                        vote.IsSlashed,
+			IsWithdrawableInCurrentEpoch:     vote.IsWithdrawableInCurrentEpoch,
+			IsActiveInCurrentEpoch:           vote.IsActiveInCurrentEpoch,
+			IsActiveInPreviousEpoch:          vote.IsActiveInPreviousEpoch,
+			IsCurrentEpochAttester:           vote.IsCurrentEpochAttester,
+			IsCurrentEpochTargetAttester:     vote.IsCurrentEpochTargetAttester,
+			IsPreviousEpochAttester:          vote.IsPreviousEpochAttester,
+			IsPreviousEpochTargetAttester:    vote.IsPreviousEpochTargetAttester,
+			IsPreviousEpochHeadAttester:      vote.IsPreviousEpochHeadAttester,
+			CurrentEpochEffectiveBalanceGwei: fmt.Sprintf("%d", vote.CurrentEpochEffectiveBalanceGwei),
+			InclusionSlot:                    fmt.Sprintf("%d", vote.InclusionSlot),
+			InclusionDistance:                fmt.Sprintf("%d", vote.InclusionDistance),
+			InactivityScore:                  fmt.Sprintf("%d", vote.InactivityScore),
+		})
+	}
+	response := &structs.GetIndividualVotesResponse{
+		IndividualVotes: v,
+	}
+	httputil.WriteJson(w, response)
 }
